+++ src/modules/see_know/tests.rs
@@ -92,7 +92,7 @@ use crate::util::key_pool::KeyStatus;
             "15551234567",
             "scan",
             "search",
-            "see-know.icu:test",
+            "see-know.ru:test",
             &mut seen,
             &mut result,
         );
@@ -163,7 +163,7 @@ use crate::util::key_pool::KeyStatus;
             "ali.kareem",
             "scan",
             "search",
-            "see-know.icu:test",
+            "see-know.ru:test",
             &mut seen,
             &mut result,
         );
@@ -183,7 +183,7 @@ use crate::util::key_pool::KeyStatus;
             "x",
             "scan",
             "search",
-            "see-know.icu:test",
+            "see-know.ru:test",
             &mut seen,
             &mut result,
         );
@@ -214,7 +214,7 @@ use crate::util::key_pool::KeyStatus;
             "Ali Kareem",
             "scan",
             "search",
-            "see-know.icu:test",
+            "see-know.ru:test",
             &mut seen,
             &mut result,
         );
@@ -264,7 +264,7 @@ use crate::util::key_pool::KeyStatus;
             "Ali Kareem",
             "scan",
             "search",
-            "see-know.icu:test",
+            "see-know.ru:test",
             &mut seen,
             &mut result,
         );
@@ -305,7 +305,7 @@ use crate::util::key_pool::KeyStatus;
             "Ali Kareem",
             "scan",
             "search",
-            "see-know.icu:test",
+            "see-know.ru:test",
             &mut seen,
             &mut result,
         );
@@ -337,7 +337,7 @@ use crate::util::key_pool::KeyStatus;
             "Ali Kareem",
             "scan",
             "search",
-            "see-know.icu:test",
+            "see-know.ru:test",
             &mut seen,
             &mut result,
         );
@@ -373,7 +373,7 @@ use crate::util::key_pool::KeyStatus;
             "victim_login",
             "scan",
             "stealer",
-            "see-know.icu:test",
+            "see-know.ru:test",
             &mut seen,
             &mut result,
         );
@@ -428,7 +428,7 @@ use crate::util::key_pool::KeyStatus;
             "victim_login",
             "scan",
             "stealer",
-            "see-know.icu:test",
+            "see-know.ru:test",
             &mut seen,
             &mut result,
         );
@@ -462,7 +462,7 @@ use crate::util::key_pool::KeyStatus;
             "victim@example.com",
             "scan",
             "search",
-            "see-know.icu:test",
+            "see-know.ru:test",
             &mut seen,
             &mut result,
         );
@@ -500,7 +500,7 @@ use crate::util::key_pool::KeyStatus;
             "x@y.com",
             "scan",
             "search",
-            "see-know.icu:test",
+            "see-know.ru:test",
             &mut seen2,
             &mut result2,
         );
@@ -542,7 +542,7 @@ use crate::util::key_pool::KeyStatus;
             "x",
             "scan",
             "search",
-            "see-know.icu:test",
+            "see-know.ru:test",
             &mut seen,
             &mut result,
         );
@@ -802,7 +802,7 @@ use crate::util::key_pool::KeyStatus;
         let before = result.entities.len();
         resolve_identity_pivots(
             "key",
-            "see-know.icu:test",
+            "see-know.ru:test",
             "seed",
             "t",
             &mut seen,